Functional but could use improvement brewtracker for integrating tilt hydrometer and brewfather

+++ pkg/brewtracker/config.go
@@ -8,7 +8,7 @@ import (
 )
 
 type ConfigPrometheus struct {
-	Port uint16 `mapstructure:"port"`
+	Port int `mapstructure:"port"`
 }
 
 type Config struct {
